Move and refactor `TopLevelErrorGeneratorTest` to not require `RustCodegenPlugin`
